requirements.txt
# Core Framework
langchain==0.1.0
langgraph==0.0.20
langchain-community==0.0.13

# LLM Providers (Ollama client)
ollama==0.1.6

# Web Crawling & Scraping
playwright==1.40.0
scrapy==2.11.0
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
requests==2.31.0
httpx==0.25.2
selenium==4.16.0
pytest
bs4
python-dateutil
jsonschema
pdfminer.six==20221105
pdfplumber==0.10.3
pdf2image
pytesseract
aiohttp

# Async & Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Vector Store & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.2
# torch==2.1.2
faiss-cpu==1.7.4

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.1.0
mammoth==1.6.0
# removed textract to avoid pdfminer conflict

# Matching Algorithms
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Web Framework (Dashboard)
fastapi==0.108.0
uvicorn[standard]==0.25.0
jinja2==3.1.2
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
tenacity==8.2.3

# Rate Limiting & Robots
robotexclusionrulesparser==1.7.1
ratelimit==2.2.1

# Logging & Monitoring
loguru==0.7.2
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio
pytest-cov

# Data Export
openpyxl==3.1.2
xlsxwriter==3.1.9
